clean up pro example viewer
