--- packages/xy-shared/components/pro/ProExampleViewer/types.ts
@@ -1,16 +1,0 @@
-export type ProExampleVariant = {
-  id: string;
-  label: string;
-};
-
-export type ProExampleConfig = {
-  id: string;
-  name?: string;
-  description?: string;
-  framework: string;
-  hidden?: boolean;
-  variants?: ProExampleVariant[];
-  free?: boolean;
-  type?: 'template' | 'example';
-  previewUrl: string;
-};
